+++ CHANGES.rst
@@ -13,6 +13,10 @@ Changes
   in release 0.4.0. Removed the option from man page.
 * Added option to PNG serializer to provide more than two colors. Each module
   type may have it's own color.
+* Added support for EPC QR Codes.
+* Fixed bug in ``helpers.make_vcard_data`` function (the "source" URL was not
+  used, but the usual URL was added to the SOURCE field)
+* Better test coverage for the ``segno.helpers`` module
 
 
 0.3.3 -- 2019-12-29